Little bit of tidy up of install code. Moved fd files into fault directory and made them fault params instead of sim params.
