+++ scripts/install.py
@@ -10,21 +10,16 @@ python ~/slurm_gm_workflow/scripts/install.py --user_root ~/Albury_newman
 import os
 import sys
 import glob
-import shutil
 import fnmatch
 import configparser
 import argparse
 from datetime import datetime
 
-import yaml
-from qcore import utils
-from qcore import simulation_structure
-from scripts.management import create_mgmt_db
-
-# TODO: namespacing
-from shared_workflow import shared
 import shared_workflow.shared_defaults as defaults
-
+from qcore import utils, simulation_structure
+from scripts.management import create_mgmt_db
+from shared_workflow.install_shared import install_simulation, dump_all_yamls
+from shared_workflow import shared
 
 def q_accept_custom_rupmodel():
     shared.show_horizontal_line()
@@ -207,26 +202,7 @@ def q_get_run_name(HH, srf_selected, v_mod_ver, emod3d_version):
     return yes, run_name
 
 
-def q_statcords_convert():
-    shared.show_horizontal_line()
-    print(
-        "Do you want the statcords and fd_statlist "
-        "to be automatically generated? (Recommended)"
-    )
-    shared.show_horizontal_line()
-    return shared.show_yes_no_question()
-
-
-def q_model_params():
-    shared.show_horizontal_line()
-    print(
-        "Do you want the model params " "to be automatically generated? (Recommended)"
-    )
-    shared.show_horizontal_line()
-    return shared.show_yes_no_question()
-
-
-def q_final_confirm(run_name, yes_statcords, yes_model_params):
+def q_final_confirm(run_name, yes_statcords):
     print("")
     shared.show_horizontal_line(c="*")
     print("S U M M A R Y")
@@ -239,279 +215,6 @@ def q_final_confirm(run_name, yes_statcords, yes_model_params):
     shared.show_horizontal_line(c="*")
     print("Do you wish to proceed?")
     return shared.show_yes_no_question()
-
-
-# =======install_bb_related==============================================================================================
-def q0():
-    shared.show_horizontal_line()
-    print(
-        "Do you want site-specific computation? "
-        "(To use a universal 1D profile, Select 'No')"
-    )
-    shared.show_horizontal_line()
-    return shared.show_yes_no_question()
-
-
-def q1_generic(v_mod_1d_dir):
-    shared.show_horizontal_line()
-    print("Select one of 1D Velocity models (from %s)" % v_mod_1d_dir)
-    shared.show_horizontal_line()
-
-    v_mod_1d_options = glob.glob(os.path.join(v_mod_1d_dir, "*.1d"))
-    v_mod_1d_options.sort()
-
-    v_mod_1d_selected = shared.show_multiple_choice(v_mod_1d_options)
-    print(v_mod_1d_selected)  # full path
-    v_mod_1d_name = os.path.basename(v_mod_1d_selected).replace(".1d", "")
-
-    return v_mod_1d_name, v_mod_1d_selected
-
-
-def install_bb(
-    stat_file,
-    root_dict,
-    v1d_dir=defaults.vel_mod_dir,
-    v1d_full_path=None,
-    site_v1d_dir=None,
-    hf_stat_vs_ref=None,
-):
-    shared.show_horizontal_line(c="*")
-    print(" " * 37 + "EMOD3D HF/BB Preparation Ver.slurm")
-    shared.show_horizontal_line(c="*")
-    root_dict["bb"] = {}
-    if v1d_full_path is not None:
-        v_mod_1d_selected = v1d_full_path
-        root_dict["bb"]["site_specific"] = False
-        root_dict["v_mod_1d_name"] = v_mod_1d_selected
-
-    # TODO:add in logic for site specific as well, if the user provided as args
-    elif site_v1d_dir is not None and hf_stat_vs_ref is not None:
-        v_mod_1d_path, hf_stat_vs_ref = shared.get_site_specific_path(
-            os.path.dirname(stat_file),
-            hf_stat_vs_ref=hf_stat_vs_ref,
-            v1d_mod_dir=site_v1d_dir,
-        )
-        root_dict["bb"]["site_specific"] = True
-        root_dict["v_mod_1d_name"] = v_mod_1d_path
-        root_dict["hf_stat_vs_ref"] = hf_stat_vs_ref
-        root_dict["bb"]["rand_reset"] = True
-    else:
-        is_site_specific_id = q0()
-        if is_site_specific_id:
-            v_mod_1d_path, hf_stat_vs_ref = shared.get_site_specific_path(
-                os.path.dirname(stat_file)
-            )
-            root_dict["bb"]["site_specific"] = True
-            root_dict["v_mod_1d_name"] = v_mod_1d_path
-            root_dict["hf_stat_vs_ref"] = hf_stat_vs_ref
-            root_dict["bb"]["rand_reset"] = True
-        else:
-            v_mod_1d_name, v_mod_1d_selected = q1_generic(v1d_dir)
-            root_dict["bb"]["site_specific"] = False
-            root_dict["v_mod_1d_name"] = v_mod_1d_selected
-
-
-def action(
-    version,
-    sim_dir,
-    event_name,
-    run_name,
-    run_dir,
-    vel_mod_dir,
-    srf_file,
-    stoch_file,
-    params_vel_path,
-    stat_file_path,
-    vs30_file_path,
-    vs30ref_file_path,
-    MODEL_LAT,
-    MODEL_LON,
-    MODEL_ROT,
-    hh,
-    nx,
-    ny,
-    nz,
-    sufx,
-    sim_duration,
-    vel_mod_params_dir,
-    yes_statcords,
-    yes_model_params,
-    fault_yaml_path,
-    root_yaml_path,
-    v1d_dir=defaults.vel_mod_dir,
-    user_root=defaults.user_root,
-    stat_dir=defaults.stat_dir,
-    site_v1d_dir=None,
-    hf_stat_vs_ref=None,
-    v1d_full_path=None,
-    sim_params_file="",
-):
-    lf_sim_root_dir = os.path.join(sim_dir, "LF")
-    hf_dir = os.path.join(sim_dir, "HF")
-    bb_dir = os.path.join(sim_dir, "BB")
-    dir_list = [sim_dir, lf_sim_root_dir, hf_dir, bb_dir]
-    version = str(version)
-    if not os.path.isdir(user_root):
-        dir_list.insert(0, user_root)
-
-    shared.verify_user_dirs(dir_list)
-
-    for filename in glob.glob(os.path.join(defaults.recipe_dir, "*.*")):
-        if filename == "README.md":
-            continue
-        shutil.copy(filename, sim_dir)
-
-        # TODO: the next two lines are two files for old post-processing
-        # shutil.copy(os.path.join(gmsa_dir,"parametersStation.py"),sim_dir)
-        # shutil.copy(os.path.join(gmsa_dir,"runPostProcessStation.ll"),sim_dir)
-    #    exe('ln -s %s/submit_emod3d.py %s'%(bin_process_dir,sim_dir))
-    shutil.copy(os.path.join(defaults.workflow_root, "version"), sim_dir)
-    shutil.copy(os.path.join(defaults.bin_process_dir, "submit.sh"), sim_dir)
-
-    if not yes_model_params:
-        print("Generation of model params has been skipped.")
-        print("Re-directing related params to files under %s" % vel_mod_dir)
-        vel_mod_params_dir = vel_mod_dir
-
-    template_path = os.path.join(defaults.recipe_dir, "gmsim", version)
-    root_params_dict = utils.load_yaml(
-        os.path.join(template_path, "root_defaults.yaml")
-    )
-    fault_params_dict = {}
-    sim_params_dict = {}
-    vm_params_dict = {}
-
-    sim_params_dict["fault_yaml_path"] = fault_yaml_path
-    fault_params_dict["root_yaml_path"] = root_yaml_path
-
-    sim_params_dict["run_name"] = run_name
-    # select during install
-    root_params_dict["version"] = version
-
-    root_params_dict["stat_file"] = stat_file_path
-    # potential remove
-    sim_params_dict["user_root"] = user_root
-    sim_params_dict["run_dir"] = run_dir
-
-    sim_params_dict["sim_dir"] = sim_dir
-
-    sim_params_dict["srf_file"] = srf_file
-    fault_params_dict["vel_mod_dir"] = vel_mod_dir
-    sim_params_dict["params_vel"] = params_vel_path
-    sim_params_dict["sim_duration"] = sim_duration
-
-    vm_params_dict["MODEL_LAT"] = MODEL_LAT
-    vm_params_dict["MODEL_LON"] = MODEL_LON
-    vm_params_dict["MODEL_ROT"] = MODEL_ROT
-    vm_params_dict["hh"] = hh
-    vm_params_dict["nx"] = nx
-    vm_params_dict["nz"] = nz
-    vm_params_dict["ny"] = ny
-
-    vm_params_dict["sufx"] = sufx
-    vm_params_dict["GRIDFILE"] = os.path.join(vel_mod_params_dir, "gridfile%s" % sufx)
-    vm_params_dict["GRIDOUT"] = os.path.join(vel_mod_params_dir, "gridout%s" % sufx)
-    vm_params_dict["MODEL_COORDS"] = os.path.join(
-        vel_mod_params_dir, "model_coords%s" % sufx
-    )
-    vm_params_dict["MODEL_PARAMS"] = os.path.join(
-        vel_mod_params_dir, "model_params%s" % sufx
-    )
-    vm_params_dict["MODEL_BOUNDS"] = os.path.join(
-        vel_mod_params_dir, "model_bounds%s" % sufx
-    )
-
-    sim_params_dict["hf"] = {}
-    sim_params_dict["hf"]["hf_slip"] = stoch_file
-
-    sim_params_dict["bb"] = {}
-
-    # potential remove
-    sim_params_dict["emod3d"] = {}
-    sim_params_dict["emod3d"]["n_proc"] = 512
-    if stat_file_path == "":
-        # stat_path seems to empty, assigning all related value to latest_ll
-        print("stat_file_path is not specified.")
-        print("Using %s" % defaults.latest_ll)
-        run_stat_dir = os.path.join(stat_dir, event_name)
-        stat_file_path = os.path.join(run_stat_dir, event_name + ".ll")
-        vs30_file_path = os.path.join(run_stat_dir, event_name + ".vs30")
-        vs30ref_file_path = os.path.join(run_stat_dir, event_name + ".vs30ref")
-
-        # creating sub-folder for run_name
-        # check if folder already exist
-        if not os.path.isdir(run_stat_dir):
-            # folder not exist, creating
-            os.mkdir(run_stat_dir)
-
-            # making symbolic link to latest_ll
-            cmd = "ln -s %s %s" % (
-                os.path.join(defaults.latest_ll_dir, defaults.latest_ll + ".ll"),
-                stat_file_path,
-            )
-            shared.exe(cmd)
-
-            # making symbolic link to lastest_ll.vs30 and .vs30ref
-            cmd = "ln -s %s %s" % (
-                os.path.join(defaults.latest_ll_dir, defaults.latest_ll + ".vs30"),
-                vs30_file_path,
-            )
-            shared.exe(cmd)
-
-            cmd = "ln -s %s %s" % (
-                os.path.join(defaults.latest_ll_dir, defaults.latest_ll + ".vs30ref"),
-                vs30ref_file_path,
-            )
-            shared.exe(cmd)
-
-    root_params_dict["stat_vs_est"] = vs30_file_path
-    root_params_dict["stat_vs_ref"] = vs30ref_file_path
-
-    if stat_file_path is not None:
-        sim_params_dict["stat_file"] = stat_file_path
-
-    # if yes_model_params:
-    #     print "Producing model params. It may take a minute or two"
-    #     from gen_coords import gen_coords
-    #     gen_coords()
-    #     print "Done"
-    shared.show_horizontal_line(c="*")
-
-    if yes_statcords:
-        print("Producing statcords and FD_STATLIST. It may take a minute or two")
-
-        # Create Stat_cord & statList
-        import scripts.statlist2gp as sl2gp
-
-        fd_statcords, fd_statlist = sl2gp.main(
-            sim_params_dict, vm_params_dict, stat_file=stat_file_path
-        )
-        print("Done")
-        sim_params_dict["stat_coords"] = fd_statcords
-        sim_params_dict["FD_STATLIST"] = fd_statlist
-    else:
-        print(
-            "Generation of statcords is skipped. "
-            "You need to fix params_base.py manually"
-        )
-    print("installing bb")
-    install_bb(
-        stat_file_path,
-        root_params_dict,
-        v1d_dir=v1d_dir,
-        v1d_full_path=v1d_full_path,
-        site_v1d_dir=site_v1d_dir,
-        hf_stat_vs_ref=hf_stat_vs_ref,
-    )
-    print("installing bb finished")
-
-    if sim_params_file and os.path.isfile(sim_params_file):
-        with open(sim_params_file) as spf:
-            extra_sims_params = yaml.load(spf)
-        for key in extra_sims_params.keys():
-            sim_params_dict.update({key: extra_sims_params[key]})
-
-    return root_params_dict, fault_params_dict, sim_params_dict, vm_params_dict
 
 
 def show_instruction(sim_dir):
@@ -557,17 +260,6 @@ def q_wallclock():
     print("Do you wish to modify the wallclock time limit?")
     shared.show_horizontal_line()
     return shared.show_yes_no_question()
-
-
-def dump_all_yamls(
-    sim_dir, root_params_dict, fault_params_dict, sim_params_dict, vm_params_dict
-):
-    utils.dump_yaml(sim_params_dict, os.path.join(sim_dir, "sim_params.yaml"))
-    utils.dump_yaml(fault_params_dict, sim_params_dict["fault_yaml_path"])
-    utils.dump_yaml(root_params_dict, fault_params_dict["root_yaml_path"])
-    utils.dump_yaml(
-        vm_params_dict, os.path.join(fault_params_dict["vel_mod_dir"], "vm_params.yaml")
-    )
 
 
 def main_local(args):
@@ -632,7 +324,7 @@ def main_local(args):
     fault_yaml_path = simulation_structure.get_fault_yaml_path(sim_dir)
     root_yaml_path = simulation_structure.get_root_yaml_path(sim_dir)
 
-    root_params_dict, fault_params_dict, sim_params_dict, vm_params_dict = action(
+    root_params_dict, fault_params_dict, sim_params_dict, vm_params_dict = install_simulation(
         args.version,
         sim_dir,
         event_name,
@@ -716,7 +408,7 @@ def main_remote(cfg, args):
 
     # TODO action will return 4 params dict and they will be dumped into yamls.
     # TODO to implement when install_manual is merged
-    root_params_dict, fault_params_dict, sim_params_dict, vm_params_dict = action(
+    root_params_dict, fault_params_dict, sim_params_dict, vm_params_dict = install_simulation(
         args.version,
         sim_dir,
         event_name,
